Move hypersync health check into a dedicated integration test binary
Promote the health check from an inline #[cfg(test)] module to
tests/hypersync_health.rs so CI can select it with `--test hypersync_health`
instead of a brittle module-path substring filter. A non-matching substring
filter silently runs zero tests and still exits 0; selecting the test binary
by name cannot.
