+++ .github/workflows/build_and_verify.yml
@@ -156,7 +156,7 @@ jobs:
       - uses: ./.github/actions/cargo-cache
 
       - name: Hypersync Health Test
-        run: cargo test --features hypersync_health config_parsing::hypersync_endpoints::hypersync_health
+        run: cargo test --features hypersync_health --test hypersync_health
 
   # Template tests - verify envio templates can be initialized and built
   template-tests:
